Initial commit of Lights Out!
Lights Out! is a copy of an electronic game that I used to have where
pressing a button will toggle that button plus all of the buttons directly
surrounding it (not diagonals). The goal is to get all of the buttons
turned off. This is the intial commit of the files with a playable version
of the game with a single starting state and 3 grid sizes to choose from.